Refactor GMMRegression to a separate gmr/sklearn.py module
